+++ build.gradle
@@ -21,7 +21,7 @@ configurations {
 
 repositories {
     mavenCentral()
-    maven { url 'https://dl.bintray.com/jaypark3749/maven/'}
+    maven { url 'https://dl.bintray.com/jaypark3749/maven/' }
 }
 
 ext {
@@ -48,9 +48,9 @@ dependencies {
     testImplementation('org.springframework.boot:spring-boot-starter-test') {
         exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
     }
-    testImplementation 'org.springframework.security:spring-security-test'
     testImplementation 'org.springframework.restdocs:spring-restdocs-mockmvc'
     testImplementation 'org.springframework.security:spring-security-test'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation 'org.springframework.boot:spring-boot-starter-webflux' // webtestclient 를 위한 추가
 }
