+++ src/main/java/com/gaejangmo/apiserver/config/oauth/service/CustomOAuth2UserService.java
@@ -1,29 +1,18 @@
 package com.gaejangmo.apiserver.config.oauth.service;
 
-import com.gaejangmo.apiserver.config.oauth.dto.OAuthAttributesDto;
-import com.gaejangmo.apiserver.model.common.domain.vo.Link;
-import com.gaejangmo.apiserver.model.user.domain.User;
-import com.gaejangmo.apiserver.model.user.domain.UserRepository;
-import com.gaejangmo.apiserver.model.user.domain.vo.Email;
-import com.gaejangmo.apiserver.model.user.domain.vo.Grade;
-import com.gaejangmo.apiserver.model.user.domain.vo.Role;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserService;
 import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
-import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 
-import java.util.Collections;
-
 @Service
 public class CustomOAuth2UserService implements OAuth2UserService<OAuth2UserRequest, OAuth2User> {
-    private final UserRepository userRepository;
+    private final OAuth2Service oauthService;
 
-    public CustomOAuth2UserService(final UserRepository userRepository) {
-        this.userRepository = userRepository;
+    public CustomOAuth2UserService(final OAuth2Service oauthService) {
+        this.oauthService = oauthService;
     }
 
     @Override
@@ -35,33 +24,6 @@ public class CustomOAuth2UserService implements OAuth2UserService<OAuth2UserRequ
         String userNameAttributeName = userRequest.getClientRegistration().getProviderDetails()
                 .getUserInfoEndpoint().getUserNameAttributeName();
 
-        OAuthAttributesDto attributes = OAuthAttributesDto.of(registrationId, userNameAttributeName, oAuth2User.getAttributes());
-
-        User user = saveOrUpdate(attributes);
-
-        return new DefaultOAuth2User(
-                Collections.singleton(new SimpleGrantedAuthority(user.getRoleType())),
-                attributes.getAttributes(),
-                attributes.getNameAttributeKey()
-        );
-    }
-
-    private User saveOrUpdate(final OAuthAttributesDto attributes) {
-        User user = userRepository.findByEmail(Email.of(attributes.getEmail()))
-                .map(userEntity -> userEntity.update(attributes.getUsername(), attributes.getImageUrl()))
-                .orElseGet(() -> toEntity(attributes));
-
-
-        return userRepository.save(user);
-    }
-
-    private User toEntity(final OAuthAttributesDto attributes) {
-        return User.builder()
-                .username(attributes.getUsername())
-                .email(Email.of(attributes.getEmail()))
-                .imageUrl(Link.of(attributes.getImageUrl()))
-                .role(Role.USER)
-                .grade(Grade.GENERAL)
-                .build();
+        return oauthService.getUser(registrationId, userNameAttributeName, oAuth2User.getAttributes());
     }
 }
